Lab04/requirements.txt
-e .
aws-cdk.aws-s3>=1.6.0
aws-cdk.aws-rds>=1.6.0
